Keep the submit footer off the package's public surface
`package.json` maps `./components/*` onto every `src/components/*.tsx`, so
splitting the action row into its own file published a form-only component —
and its fixed DOM id — as a supported package entry point. Colocate it with
the form instead, which adds no new subpath and matches how the rest of the
module's test-facing helpers are already exposed.

Also sorts the new test file's imports, which is what CI's root
`biome check .` caught: the package lint script is `biome check src/` and
never sees `tests/`.
